Suggested fixes for PR#2
