--- src/adapters/vcs/git/types.ts
@@ -1,4 +1,0 @@
-export enum GitState {
-    initialized = 'initialized',
-    uninitialized = 'uninitialized',
-}
